Card-O-Matic working
Question about the draggable instead of click to add to card. When
dragging the sticker onto the card it displaces other stickers on the
card. I tried to add a class with position absolute to the
ui.draggable.draggable but it didn't work as planned. :( I would like
to gett this going eventually.